Update dashboard formatting, TPS/GAP types, and inline DB data for specific IP

+++ TradeDesktop.Application/Models/ExchangeMetrics.cs
@@ -6,7 +6,7 @@ public sealed record ExchangeMetrics(
     decimal? Ask,
     decimal? Spread,
     decimal? LatencyMs,
-    decimal? Tps,
+    float? Tps,
     string? Time,
     decimal? MaxLatMs,
     decimal? AvgLatMs,
